+++ packages/terra-core-docs/src/terra-dev-site/test/demographics-banner/BlankDemographicsBanner.test.jsx
@@ -1,4 +1,4 @@
 import React from 'react';
-import DemographicsBanner from '../../../DemographicsBanner';
+import DemographicsBanner from 'terra-demographics-banner';
 
 export default () => <DemographicsBanner id="blank-demographics-banner" />;