+++ packages/terra-demographics-banner/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Migrate docs to terra-core-docs.
+
 ## 3.56.0 - (October 8, 2021)
 
 * Changed